Major changes were made. Whole architecture were change. Some clean code properties were implemented.
