--- lib/PicassoCreation/picasso.dart
@@ -1,75 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:pomodoro2/MonetCreation/monet_variables.dart';
-import 'package:pomodoro2/PicassoCreation/picasso_variables.dart';
-import '../main.dart';
-import '../image_container.dart';
-
-class Picasso extends StatefulWidget {
-  const Picasso({super.key, required this.title, this.plannerId, this.date});
-  final String title;
-  final int? plannerId;
-  final String? date;
-
-  @override
-  State<Picasso> createState() => _PicassoState();
-}
-
-class _PicassoState extends State<Picasso> {
-  late Future<SingleChildScrollView> taskFuture;
-
-  @override
-  void initState(){
-    super.initState();
-    taskFuture = createPlanner(
-      widget.date!,
-      widget.plannerId!,
-      PicassoVariables.dateColor,
-      PicassoVariables.taskColor,
-      PicassoVariables.textColor,
-    );
-
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final double deviceWidth = MediaQuery.of(context).size.width;
-    return Scaffold(
-      backgroundColor: PicassoVariables.taskColor,
-      resizeToAvoidBottomInset: true,
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            const ImageContainer(
-              imageUrl: 'assets/images/Picasso/2.jpg',
-              imageAlignment: Alignment(0, -1),
-            ),
-            Flexible(
-                child: FutureBuilder<SingleChildScrollView>(
-                  future: taskFuture,
-                  builder: (context, snapshot) {
-                    if (snapshot.connectionState == ConnectionState.waiting) {
-                      return const CircularProgressIndicator();
-                    } else if (snapshot.hasError) {
-                      return Text('Error: ${snapshot.error}');
-                    } else {
-                      return snapshot.data ?? Container();
-                    }
-                  },
-                )
-            )
-          ],
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        backgroundColor: PicassoVariables.dateColor,
-        child: const Icon(
-            Icons.home_filled
-        ),
-        onPressed: () {
-          Navigator.pop(context);
-        },
-      ),
-    );
-  }
-}
