--- lib/common_variables.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/cupertino.dart';
-
-class Variables {
- static const double heightProportionOfImageBox = 0.25;
- static const double heightProportionOfDateContainer = 0.075;
- static const double widthOfDayInDateContainer = 0.25;
- static const double widthOfDateInDateContainer = 0.55;
- static const double heightProportionOfTaskContainer = 0.3;
- static EdgeInsets fixedEdgeInsets = const EdgeInsets.all(10.0);
- static double textSize = 17.0;
- static double inputTextSize = 14.0;
- static Color homePageColor = const Color.fromRGBO(213, 227, 250, 1.0);
-}
