+++ lib/ui/helper/common_variables.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+const double heightProportionOfImageBox = 0.25;
+const double heightProportionOfDateContainer = 0.075;
+const double widthOfDayInDateContainer = 0.25;
+const double widthOfDateInDateContainer = 0.55;
+const double heightProportionOfTaskContainer = 0.3;
+const EdgeInsets fixedEdgeInsets = EdgeInsets.all(10.0);
+const double textSize = 17.0;
+const double inputTextSize = 14.0;
+const Color homePageColor = Color.fromRGBO(213, 227, 250, 1.0);
+const IconData homeIcon = Icons.home_filled;
+const Color homeIconFillingColor = Color.fromRGBO(44, 54, 63, 1);
+const Map<String, String> artists =
+{
+  "GustavKlimt":"Gustav Klimt",
+  "OsmanHamdi":"Osman Hamdi Bey",
+  "Monet":"Monet",
+  "Picasso":"Picasso",
+  "SalvadorDali":"Salvador Dali",
+  "VanGogh":"Van Gogh"
+};
+const dismissibleColor = Color.fromRGBO(242, 245, 234, 1);
