+++ lib/ui/helper/created_planner_variables.dart
@@ -1,0 +1,2 @@
+import 'dart:ui';
+const Color floatingActionButtonBackgroundColor = Color.fromRGBO(242, 245, 234, 1);
