+++ lib/ui/helper/gustav_klimt_variables.dart
@@ -1,6 +1,4 @@
 import 'package:flutter/material.dart';
-
-
 class GustavKlimtVariables {
   static Color dateColor = const Color.fromRGBO(216, 162, 75, 1);
   static Color? taskColor = const Color.fromRGBO(241, 197, 80, 1);
