+++ lib/ui/helper/osman_hamdi_variables.dart
