+++ lib/ui/helper/picasso_variables.dart
